Move `LiveDataExtensions` to module data to reuse them
